Remove blob hashes from publication
The PublicationHook can check the hashes if desired, and/or return
them in the auxData. The number of blobs can also be passed in the
data field.

The intention is to explicitly remove any data availability requirements.
This enables the possibility for publication hooks to add non-validated
publications to the queue