Ensure evictProver is based on an unproven checkpoint
The CheckpointTracker only saves a hash of the last proven checkpoint.
I add a parameter so the caller can provide the actual checkpoint to be
checked against the hash. I did not update the tests, because I think we
should actually update the CheckpointTracker to save the whole checkpoint,
but I did not want to edit it until the merge conflicts are resolved.