Use payPublicationFee transition logic in getCurrentFees
The first publication of every new period will call `getCurrentFees` before the currentPeriodId
is updated, but will pay the fees after it is updated. This means it will pay the wrong fee.
In the case the fee goes up, the publication will fail unless the proposer already has
enough of a deposit to make up the difference