Create mechanism to finalise an old period
It is possible for a period to end without the prover getting their stake.
This could occur if they forget to include nextPublicationHeaderBytes but also:
- the last publication for a while might occur in their proving window, so they
  cannot wait for another publication, or
- they might prove some publications during their window before they realise
  that there will be no new publications in their window.

In either case, we want a mechanism to recognise that the prover has completed
their task but has not receive their stake. This PR clears the fees and stake
when the funds are distributed so any non-zero values are still to be distributed.