Do not finalize on proveClosedPeriod. Temporarily disable another test.
This simplifies the implementation and does not change the economics
in the expected case where a single proof is used to finalise all
publications in the period.

It is possible for someone to prove only some of the publications.
They will receive the publication fee but not the stake, and implicitly
increase the burn percentage.