Remove the requirement that percentages cannot round to zero
I want to avoid the possibility that the burned stake gets large
enough that it is impossible to call proveClosedPeriod.

However, to ensure minUndercutPercentage does not round to zero,
replace it with maxBidPercentage. This ensures the bid keeps getting
smaller until it is zero.

It is possible for provers to keep outbidding each other at zero,
but I don't think that's a real problem. We could use introduce a
require(_currentPeriod.fee > 0) if necessary.