Bug fix: claimProvingVacancy should only start next block
There were two problems.

Firstly, the vacant period did not have an end timestamp, which is used to
detect the start of the next period in prove(). This means it was
possible to make a proof and claim rewards for publications that occured
before the new period. This would be fine if it only applied to the vacant period
but it would apply to the previous period as well.

By immediately creating a new period, it is possible for some publications
to occur before and after the claimProvingVacancy call within the same block.
They will be charged different fees but would be attributed to the same period
(the vacant period) for proving purposes.