Allow proofs for period 0
We create a new period in the constructor, and I originally thought that
implied we do not need to prove publications in period 0. However, the
_claimProvingVacancy call implies any publication created in the deployment
block is assigned to period 0. Moreover, under all versions in this PR,
any publication that occured before deployment is assigned to period 0.
I now think we should account for this because the ProverManager might be
deployed on a shared publication feed that pre-exists the CheckpointTracker.
I also don't think we should rely on the CheckpointTracker being able to
prove things before the prover manager is assigned, since this puts a
specific constraint on the deployment mechanism.