Rename storage commitment error message and change comment
The previous comment was ambiguous about whether we support full state roots.
I think `StateProofNotSupported` as an error is unclear, and might actually state the opposite of the error.

For context
- the `LibSignal` library attempts to prove a storage location inside a root.
- we expect this to be a state root. The proof is multi-step: first prove the account's storage root within the state proof, then prove the storage location against the storage root.
- the library also supports skipping step 1 and treating the root as a storage root directly. The signal service does not.

At some point we will have to generalise the commitment and storage proof mechanism to support block hashes as commitments (which is natural for rollups) and arbitrary state for appchains. In the mean time we are being opinionated that the commitment is a state root.