Create test for verifySignal
The previous SignalService.t.sol had hardcoded proof values, which implies one
of the offchain/*.rs files needed to be run ahead of time. Using this pattern,
that rust script now generates a sample proof that can be imported directly into
the test, so we don't need to have multiple forks occuring within the test.

The codebase is inconsistent. After applying the same transformation to the bridge
tests, I will clean up the obsolete functions and scripts